compiler script started
