+++ outer.forth
@@ -1,0 +1,1 @@
+:
